fix(scenario-file): Suppress nullpo while attempting to deserialize BlocKBean

+++ ScenamaticaModels/src/main/java/net/kunmc/lab/scenamatica/interfaces/scenariofile/misc/BlockBean.java
@@ -58,7 +58,7 @@ public interface BlockBean
      *
      * @return ブロックの明るさ
      */
-    byte getLightLevel();
+    int getLightLevel();
 
     /**
      * ブロックのバイオームを取得します。
